refactor(ui): put some features in folders
