+++ packages/extension/src/ui/components/Account/AccountListItem.tsx
@@ -7,6 +7,7 @@ import styled, { css } from "styled-components"
 import useSWR from "swr"
 
 import { Account } from "../../Account"
+import { recover } from "../../features/recovery/recovery.service"
 import { useNetwork } from "../../hooks/useNetworks"
 import { routes } from "../../routes"
 import { useAccount } from "../../states/account"
@@ -19,7 +20,6 @@ import { makeClickable } from "../../utils/a11y"
 import { AccountStatus, getAccountImageUrl } from "../../utils/accounts"
 import { truncateAddress } from "../../utils/addresses"
 import { deleteAccount } from "../../utils/messaging"
-import { recover } from "../../utils/recovery"
 import { checkIfUpgradeAvailable } from "../../utils/upgrade"
 import { NetworkStatusWrapper } from "../NetworkSwitcher"
 import { TransactionStatusIndicator } from "../StatusIndicator"
