+++ packages/extension/src/ui/components/Account/AccountTokens.tsx
@@ -5,6 +5,8 @@ import styled from "styled-components"
 import useSWR from "swr"
 
 import { Account } from "../../Account"
+import { useBackupRequired } from "../../features/recovery/backupDownload.state"
+import { RecoveryBanner } from "../../features/recovery/RecoveryBanner"
 import { useAccountStatus } from "../../hooks/useAccountStatus"
 import { useNetwork } from "../../hooks/useNetworks"
 import { routes } from "../../routes"
@@ -14,11 +16,9 @@ import {
 } from "../../states/accountMetadata"
 import { useAccountTransactions } from "../../states/accountTransactions"
 import { useAppState } from "../../states/app"
-import { useBackupRequired } from "../../states/backupDownload"
 import { makeClickable } from "../../utils/a11y"
 import { connectAccount } from "../../utils/accounts"
 import { checkIfUpgradeAvailable } from "../../utils/upgrade"
-import { RecoveryBanner } from "../RecoveryBanner"
 import { Spinner } from "../Spinner"
 import { AddTokenIconButton, TokenTitle, TokenWrapper } from "../Token"
 import { AccountSubHeader } from "./AccountSubheader"
