+++ packages/extension/src/ui/features/recovery/SeedRecoveryScreen.tsx
@@ -1,13 +1,13 @@
 import { FC, useMemo, useRef, useState } from "react"
 
-import { BackButton } from "../components/BackButton"
-import { TextArea } from "../components/InputText"
-import { IconBar } from "../components/Recovery/IconBar"
-import { A, FormError, P } from "../components/Typography"
-import { useCustomNavigate } from "../hooks/useCustomNavigate"
-import { routes } from "../routes"
-import { validateAndSetSeedPhrase } from "../states/seedRecover"
-import { ConfirmScreen } from "./ConfirmScreen"
+import { BackButton } from "../../components/BackButton"
+import { TextArea } from "../../components/InputText"
+import { A, FormError, P } from "../../components/Typography"
+import { useCustomNavigate } from "../../hooks/useCustomNavigate"
+import { routes } from "../../routes"
+import { ConfirmScreen } from "../../screens/ConfirmScreen"
+import { IconBar } from "./IconBar"
+import { validateAndSetSeedPhrase } from "./seedRecover.state"
 
 export const SeedRecoveryScreen: FC = () => {
   const textAreaElement = useRef<HTMLTextAreaElement>(null)
