+++ packages/extension/src/ui/features/recovery/recovery.service.ts
@@ -1,10 +1,10 @@
-import { accountsOnNetwork, defaultNetwork } from "../../shared/networks"
-import { Account } from "../Account"
-import { routes } from "../routes"
-import { useAccount } from "../states/account"
-import { setDefaultAccountNames } from "../states/accountMetadata"
-import { useAppState } from "../states/app"
-import { getAccounts, getLastSelectedAccount } from "./messaging"
+import { accountsOnNetwork, defaultNetwork } from "../../../shared/networks"
+import { Account } from "../../Account"
+import { routes } from "../../routes"
+import { useAccount } from "../../states/account"
+import { setDefaultAccountNames } from "../../states/accountMetadata"
+import { useAppState } from "../../states/app"
+import { getAccounts, getLastSelectedAccount } from "../../utils/messaging"
 
 interface RecoveryOptions {
   networkId?: string
