+++ packages/extension/src/ui/screens/BackupDownloadScreen.tsx
@@ -7,8 +7,8 @@ import { BackButton } from "../components/BackButton"
 import { Button } from "../components/Button"
 import { Header } from "../components/Header"
 import { H2, P } from "../components/Typography"
+import { useBackupRequired } from "../features/recovery/backupDownload.state"
 import { routes } from "../routes"
-import { useBackupRequired } from "../states/backupDownload"
 
 const DownloadButton = styled(Button)`
   margin-top: auto;
