+++ packages/extension/src/ui/screens/LockScreen.tsx
@@ -8,12 +8,12 @@ import { Button } from "../components/Button"
 import { Greetings, GreetingsWrapper } from "../components/Greetings"
 import { InputText } from "../components/InputText"
 import { A, FormError, P } from "../components/Typography"
+import { recover } from "../features/recovery/recovery.service"
+import { validatePassword } from "../features/recovery/seedRecover.state"
 import { routes } from "../routes"
 import { useAppState } from "../states/app"
-import { validatePassword } from "../states/seedRecover"
 import { makeClickable } from "../utils/a11y"
 import { startSession } from "../utils/messaging"
-import { recover } from "../utils/recovery"
 import { StickyGroup } from "./ConfirmScreen"
 
 const LockScreenWrapper = styled.div`
