+++ packages/extension/src/ui/screens/NewWalletScreen.tsx
@@ -8,12 +8,12 @@ import { Button } from "../components/Button"
 import { Header } from "../components/Header"
 import { InputText } from "../components/InputText"
 import { FormError, H2, P } from "../components/Typography"
+import { recover } from "../features/recovery/recovery.service"
+import { validatePassword } from "../features/recovery/seedRecover.state"
 import { routes } from "../routes"
 import { useAccount } from "../states/account"
 import { useAppState } from "../states/app"
-import { validatePassword } from "../states/seedRecover"
 import { connectAccount, deployAccount } from "../utils/accounts"
-import { recover } from "../utils/recovery"
 import { StickyGroup } from "./ConfirmScreen"
 
 const Container = styled.div`
